Tidy up, move registartion to its own file
